requirements.txt
requests
beautifulsoup4
aiohttp
lxml
feedparser
python-dotenv
